requirements.txt
numpy
schedule
pandas==1.5.2
requests
python-telegram-bot==13.5
matplotlib
seaborn
ordered-set
ta-lib==0.4.32
sqlalchemy==1.4.39
quantstats==0.0.59
scikit-learn==1.2.2
prettytable==3.6.0
python-dateutil==2.8.2
mplfinance
tensorflow==2.12.0
fastapi
uvicorn
anthropic
python-dotenv
mcp

